Added pip install approach for python-ev3dev
This tutorial has a step to install the python-ev3dev source code on your host computer just for the sake of auto complete.  It won't magically make your computer start driving motors, but the code completion is SO handy.  The steps originally shown in the tutorial didn't work for me, but the standard pip install worked fine.  I left the original instructions too just in case they help someone, but I see no reason not to use pip for this task.