updated ALT-F2 option with HDMI
Update to PR to make clear that ALT-F2 boot to login prompt with attached HDMI screen is an alternative way to establish first connection to pi, in addition to Ethernet.  Also added parenthetic remark with external link to make clear that a display for Brickman can be added to BrickPi.