Add hints for RPi users using HDMI display (no Brickman)
* link to connmanctl help and ALT-F2 for pi users

Added hyperlink to Connman wiki that gives basic commands to establish wifi/bluetooth connections on the pi.  In same paragraph, for pi users booting with HDMI screen added note about ALT-F2 to get login prompt.

* updated ALT-F2 option with HDMI

Update to PR to make clear that ALT-F2 boot to login prompt with attached HDMI screen is an alternative way to establish first connection to pi, in addition to Ethernet.  Also added parenthetic remark with external link to make clear that a display for Brickman can be added to BrickPi.

* updated networking. md to remove stray 'Note'

to fix language in earlier PR to change discussion of ALT-F2 and HDMI

* update networking.md 

Back to simpler version that explains ALT-F2 method to boot up with HDMI screen. Removed reference to adding a BrickPi display

* point out Ctrl_Alt-F6 and connmanctl

For pi users with hdmi: Ctrl+Alt+F6 to login and then link to page with connmanctl usage for wifi or other connections

* Use relative link for connmanctl page